+++ src/main/kotlin/calculator/domain/Expression.kt
@@ -1,4 +1,4 @@
-package domain
+package calculator.domain
 
 data class Expression(private val terms: List<Term>) {
     fun getSum(): Int = terms.sumOf { it.value }
