+++ src/main/kotlin/calculator/domain/ExpressionFactory.kt
@@ -1,4 +1,4 @@
-package domain
+package calculator.domain
 
 object ExpressionFactory {
     private val DEFAULT_SEPARATORS = "[,:]".toRegex()
