--- src/main/kotlin/io/InputView.kt
@@ -1,8 +1,0 @@
-package io
-
-object InputView {
-    fun getInput(): String? {
-        println("수식을 입력하세요.")
-        return readlnOrNull()
-    }
-}
